age analysis introduction
trouble knitting rmd to html, using pdf for now but not sure what's going on.

started some analysis with age data.  Exploring VIP filtering of wavelengths for PLS models, but with 90/10 split and no split.

Pre-processing needs to be explored more.  Also a more clear work flow needs to be established.  i.e. organize the following

1) unprocessed data containing ages
2) pre-process (savitzky-golay, but maybe others?)
3) model (PLS, PCR, MLR)
4) select variables?
5) adjust/compare model performance